working on the styling of the nav and landing page for mobile vs desktop. just doing the hamburger nav on internal pages and not the home page
