+++ src/assets/styles/Home.css
@@ -4,13 +4,30 @@
     url("../../assets/fonts/Beachfly.woff2") format("woff2");
 }
 
+/* .landing_mobile-nav-container {
+  display: flex;
+  justify-content: end;
+  align-items: center;
+  height: 4.5rem;
+  padding-right: 2rem;
+  margin-bottom: 1rem;
+  position: relative;
+  background-color: rgba(34, 61, 60, 0.75);
+  z-index: 2;
+} */
+
+/* Hide the blue transparent background for desktop nav on mobile view */
+/* .landing_main-content_right {
+  display: none;
+}
+
 .desktop-navigation {
   display: none;
 }
 
 .mobile-navigation {
   display: inline;
-}
+} */
 
 .App-container {
   position: absolute;
@@ -45,13 +62,7 @@
 
 .landing_main-content_left {
   display: flex;
-  height: 60%;
-}
-
-.landing_main-content_right {
-  display: flex;
-  justify-content: center;
-  align-items: center;
+  height: 20%;
 }
 
 .landing_main-content_left::before {
@@ -60,23 +71,8 @@
   top: 0;
   left: 0;
   width: 100%;
-  height: 60%;
+  height: 20%;
   background-color: rgba(255, 234, 151, 0.5);
-  z-index: 1;
-}
-
-.landing_main-content_right {
-  height: 40%;
-}
-
-.landing_main-content_right::before {
-  content: "";
-  position: absolute;
-  top: 0;
-  top: 60%;
-  width: 100%;
-  height: 40%;
-  background-color: rgba(34, 61, 60, 0.75);
   z-index: 1;
 }
 
@@ -84,6 +80,17 @@
   font-family: "Beachfly", serif;
   font-size: 3.5rem;
   font-weight: bolder;
+  z-index: 1;
+}
+
+.landing_main-content_right::before {
+  content: "";
+  position: absolute;
+  top: 0;
+  top: 20%;
+  width: 100%;
+  height: 80%;
+  background-color: rgba(34, 61, 60, 0.75);
   z-index: 1;
 }
 
@@ -196,6 +203,19 @@
     height: 100%;
     width: 50%;
     display: flex;
+    justify-content: center;
+    align-items: center;
+  }
+
+  .landing_main-content_right::before {
+    content: "";
+    position: absolute;
+    top: 0;
+    top: 60%;
+    width: 100%;
+    height: 40%;
+    background-color: rgba(34, 61, 60, 0.75);
+    z-index: 1;
   }
 
   .landing_main-content_left::before {
