+++ src/assets/styles/MobileNav.css
@@ -1,7 +1,3 @@
-.desktop-navigation {
-  display: none;
-}
-
 .mobile-navigation {
   background-color: rgba(0, 0, 0, 0.5);
   display: inline;
